embedding-service/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
numpy
scikit-learn
python-dotenv
